master: Added some validation to the driver load scripts.

+++ scripts/load_nvidia_driver.sh
@@ -10,6 +10,11 @@ fi
 PCI_ID="0000:01:00.0"
 VENDOR_DEVICE_ID="10de 2684"
 
+if [ $( lspci -k -s $PCI_ID | grep "driver in use" | grep nvidia | wc -l ) -eq 1 ]; then
+  echo "GPU already bound to nvidia."
+  exit 0
+fi
+
 # Unbind the GPU from vfio-pci
 echo "Unbinding the GPU from vfio-pci..."
 echo "$PCI_ID" > /sys/bus/pci/drivers/vfio-pci/unbind
@@ -19,7 +24,7 @@ echo "$VENDOR_DEVICE_ID" > /sys/bus/pci/drivers/vfio-pci/remove_id
 if [ ! -e "/sys/bus/pci/drivers/vfio-pci/$PCI_ID" ]; then
   echo "GPU successfully unbound from vfio-pci."
 else
-  echo "Failed to unbind the GPU from vfio-pci. Exiting."
+  echo "Failed to unbind the GPU from vfio-pci."
   exit 1
 fi
 
@@ -28,7 +33,7 @@ echo "Reloading NVIDIA drivers..."
 modprobe nvidia nvidia_modeset nvidia_uvm nvidia_drm
 
 if [ $? -ne 0 ]; then
-  echo "Failed to reload NVIDIA drivers. Exiting."
+  echo "Failed to reload NVIDIA drivers."
   exit 1
 else
   echo "NVIDIA drivers successfully reloaded."
